+++ changelog.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Custom filters component
 - Search favorites
 - Favorite dialog
+- Delete => favorite
 
 ## [0.0.2] - 2024-01-05